Validation dữ liệu cho các field quan trọng

+++ BrainEduBackend/src/main/java/com/brainedu/BrainEdu/service/questionService/QuestionServiceImpl.java
@@ -81,7 +81,7 @@ public class QuestionServiceImpl
                         )
 
                         .weightScore(
-                                request.getWeightScore()
+                                request.getWeightScore().floatValue()
                         )
 
                         .build();
@@ -208,7 +208,7 @@ public class QuestionServiceImpl
         );
 
         question.setWeightScore(
-                request.getWeightScore()
+                request.getWeightScore().floatValue()
         );
 
         questionRepository.save(question);
